Azure_func_fastapi_py_v1/http_trigger_fastapi/__init__.py
import azure.functions as func
from fastapi import FastAPI
from routers import tareas

app = FastAPI()
app.include_router(tareas.router)

def main(req: func.HttpRequest, context: func.Context) -> func.HttpResponse:
    return func.AsgiMiddleware(app).handle(req, context)