Azure_func_fastapi_py_v1/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
fastapi
pydantic
sqlalchemy